# openapi.yaml
openapi: 3.1.0
info:
  title: Hub Innowacji Spolecznych API
  description: |-
    REST API dla mieszkańców, pracowników i administratora.
    Publiczne: katalog, regiony, zatwierdzone pomysły, dopasowanie i asystent.
    Konto mieszkańca jest wymagane do zgłoszenia nowego pomysłu i daje dostęp
    do /api/v1/ideas?mine=true. /api/v1/me oraz raporty pracownicze wymagają
    sesji; /api/v1/admin/** wymaga roli ADMIN, a /api/v1/staff/**
    roli STAFF lub ADMIN. Zgłoszenie pomysłu powiadamia pracowników;
    zalogowany autor może odczytać odpowiedzi przez /api/v1/ideas/{id}/replies
    oraz /api/v1/notifications. Trendy są w /api/v1/admin/trends;
    administrator może edytować i publikować wiedzę.
    Rejestracja: POST /api/v1/register. Logowanie: POST /login jako formularz
    username (adres e-mail), password, _csrf; odpowiedź ustawia JSESSIONID.
    Przed POST/PUT/PATCH pobierz GET /api/v1/csrf i wyślij token w nagłówku
    X-CSRF-TOKEN. Po logowaniu pobierz token ponownie. W fetch użyj
    credentials: 'include'.
  version: v1
servers:
- url: http://localhost
  description: Generated server url
paths:
...
  /api/v1/register:
    post:
      tags:
      - account-controller
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AccountDto"
      security:
      - csrfToken: []
...
components:
  schemas:
...
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          minLength: 1
        password:
          type: string
          maxLength: 128
          minLength: 12
        displayName:
          type: string
          maxLength: 255
          minLength: 0
      required:
      - displayName
      - email
      - password
    AccountDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        displayName:
          type: string
        role:
          type: string
          enum:
          - MEMBER
          - STAFF
          - ADMIN
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      name: JSESSIONID
      in: cookie
    csrfToken:
      type: apiKey
      name: X-CSRF-TOKEN
      in: header